+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 - feature/#64: ```Lines abbreviation``` [PR from [CG Bahk](https://github.com/cgbahk)]
 - feature/#67: Added ability to count on explore page
+- fixed loc for: search, tranding
 - documentation update
 - versions bumping for: jquery, tar, fstream, js-yaml, lodash, mixin-deep
 - refactoring